Struct stemflow::ResPool

pub struct ResPool<A> where A: Access {
    // some fields omitted
}

Methods

impl<A> ResPool<A> where A: Access

fn new() -> ResPool<A>

fn new_dom(&mut self, name: String, acl: Vec<A>) -> Arc<Domain<A>>

Create a domain if it doesn't have a twin or return an existing equivalent domain (can have a different name).

fn contains_dom(&self, dom: &Arc<Domain<A>>) -> bool

fn insert_dom(&mut self, dom: Arc<Domain<A>>) -> Arc<Domain<A>>

Record a domain if it doesn't have a twin or return an existing equivalent domain (can have a different name).

fn allow(&mut self, acl: &Vec<A>) -> Option<Arc<Domain<A>>>

Get (or create) the tighter domain with all this ACL